Merge all thin-edge packages into one recipe
By merging all the recipes into one, we 1) significantly simplify the
structure of the layer, and more importantly, 2) don't fetch and compile
the source code for each recipe separately, resulting in one fetch and
one compile.

At a quick glance, the change doesn't introduce any new bugs or issues,
but to be completely sure, more thorough testing is required, so until
that is done, the changes are not to be merged.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>