Merged recipes into one once again
The changes are not to be merged because postinst scripts for tedge are
not working, i.e. stuff's entirely broken. This commit just aims to once
again merge all the packages into one recipe, integrating the changes
from the main branch.

When the work in this branch is completed, some staff should be squashed
so descriptions on these WIP commits will hopefully not end up in the
main branch this time.

This reverts commit 83b1a21cbfb3ccb6d028f449c293491fcd84c10f, reversing
changes made to 8fa0dc341f78fb1fad91d7df1a49c020281333a1.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>