Add test checking connecting to the cloud
This test suite has 5 tasks which test connecting to cumulocity cloud
and will be subsequently used to test the layer before adding any
changes. The suite shall be expanded by adding more tasks to test for
issues we had previously.

The suite is based on the test in `E2E_Tutorials` directory in thin-edge
repository, but inapplicable or broken tests were deleted.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>