Fix issue with tedge-postinsts not running
The commit fixes the issue with postinsts not running, which was
probably caused by `inherit systemd` directives in `.inc` files
overwriting systemd-related bitbake variables.

The solution makes thin-edge work, but it introduced another package to
the image, `thin-edge`, which contains just the postinsts scripts. It
would be best to improve te solution to eliminate the extraneous
package.

Signed-off-by: Marcel Guzik <marcel.guzik@inetum.com>